work in progress
Information, About me, better animations, better images
