+++ tailwind.config.ts
@@ -59,13 +59,8 @@ const config: Config = {
 					foreground: "hsl(var(--card-foreground))",
 				},
 			},
-			// borderRadius: {
-			// 	lg: "var(--radius)",
-			// 	md: "calc(var(--radius) - 2px)",
-			// 	sm: "calc(var(--radius) - 4px)",
-			// },
-			backgroundSize: {
-				"300%": "300%",
+			backdropBlur: {
+				xs: "1px",
 			},
 			keyframes: {
 				"accordion-down": {
